refactor: move utils from client.ts
