+++ .env.development.example
@@ -23,8 +23,8 @@ QUERY_LIMIT_DEFAULT=-1
 KEY=directus
 SECRET=xxx
 # https://docs.directus.io/self-hosted/config-options.html#sso-oauth2-and-openid
-AUTH_GITHUB_CLIENT_ID=gh_client_id
-AUTH_GITHUB_CLIENT_SECRET=gh_client_secret
+AUTH_GITHUB_CLIENT_ID=
+AUTH_GITHUB_CLIENT_SECRET=
 # https://docs.directus.io/self-hosted/config-options.html#admin-account
 ADMIN_EMAIL=admin@example.com
 ADMIN_PASSWORD=password
