+++ docker-compose.yml
@@ -1,4 +1,4 @@
-name: e2e-globalping-dash-directus
+name: globalping-dash-directus
 version: '3'
 services:
   mariadb:
